Moved Tiled and LDTK code into examples
Fixes #84.
